requirements.txt
aiogram>=3.7,<4.0
ollama>=0.4.0,<1.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0

